created core folder - moved api.js to core folder - created global filters in core folder - created rundomizeNumber in core folder
