+++ src/core/filters.js
@@ -1,0 +1,5 @@
+import Vue from 'vue'
+
+Vue.filter('changeFunds', value => {
+  return '$' + value.toLocaleString()
+})
