+++ src/core/rundomizeNumber.js
@@ -1,0 +1,3 @@
+export const rundomizeNumber = (num) => {
+  return Math.round(num * (0.5 + Math.random()))
+}
